Remove CMultimodalPriorUtils and move functionality into class plus tidy up
