+++ lib/maths/CMultimodalPrior.cc
@@ -8,6 +8,7 @@
 
 #include <core/CContainerPrinter.h>
 #include <core/CLogger.h>
+#include <core/CSmallVector.h>
 #include <core/CStatePersistInserter.h>
 #include <core/CStateRestoreTraverser.h>
 #include <core/Constants.h>
@@ -18,12 +19,9 @@
 #include <maths/CChecksum.h>
 #include <maths/CClusterer.h>
 #include <maths/CClustererStateSerialiser.h>
-#include <maths/CIntegration.h>
 #include <maths/CKMeansOnline1d.h>
 #include <maths/CMathsFuncs.h>
-#include <maths/CMultimodalPriorUtils.h>
 #include <maths/CNormalMeanPrecConjugate.h>
-#include <maths/COrderings.h>
 #include <maths/CPriorStateSerialiser.h>
 #include <maths/CRestoreParams.h>
 #include <maths/CSampling.h>
@@ -31,12 +29,11 @@
 #include <maths/CSolvers.h>
 #include <maths/CTools.h>
 #include <maths/CToolsDetail.h>
+#include <maths/MathsTypes.h>
 #include <maths/ProbabilityAggregators.h>
 
-#include <boost/numeric/conversion/bounds.hpp>
-#include <boost/optional.hpp>
-
 #include <cmath>
+#include <limits>
 #include <set>
 
 namespace ml {
@@ -46,8 +43,78 @@ namespace {
 
 using TSizeDoublePr = std::pair<std::size_t, double>;
 using TSizeDoublePr2Vec = core::CSmallVector<TSizeDoublePr, 2>;
+using TSizeDoublePr5Vec = core::CSmallVector<TSizeDoublePr, 5>;
 using TDoubleDoublePr = std::pair<double, double>;
 using TSizeSet = std::set<std::size_t>;
+using TDouble1Vec = core::CSmallVector<double, 1>;
+using maths_t::TDoubleWeightsAry;
+using maths_t::TDoubleWeightsAry1Vec;
+using TMeanAccumulator = CBasicStatistics::SSampleMean<double>::TAccumulator;
+
+//! \brief Wrapper to call the -log(c.d.f) of a prior object.
+class CMinusLogJointCdf {
+public:
+    template<typename T>
+    bool operator()(const T& prior,
+                    const TDouble1Vec& samples,
+                    const TDoubleWeightsAry1Vec& weights,
+                    double& lowerBound,
+                    double& upperBound) const {
+        return prior->minusLogJointCdf(samples, weights, lowerBound, upperBound);
+    }
+};
+
+//! \brief Wrapper to call the log(1 - c.d.f) of a prior object.
+class CMinusLogJointCdfComplement {
+public:
+    template<typename T>
+    bool operator()(const T& prior,
+                    const TDouble1Vec& samples,
+                    const TDoubleWeightsAry1Vec& weights,
+                    double& lowerBound,
+                    double& upperBound) const {
+        return prior->minusLogJointCdfComplement(samples, weights, lowerBound, upperBound);
+    }
+};
+
+//! \brief Wrapper of CMultimodalPrior::minusLogJointCdf function
+//! for use with our solver.
+class CLogCdf {
+public:
+    using result_type = double;
+
+    enum EStyle { E_Lower, E_Upper, E_Mean };
+
+public:
+    CLogCdf(EStyle style, const CMultimodalPrior& prior, const TDoubleWeightsAry& weights)
+        : m_Style(style), m_Prior(&prior), m_Weights(1, weights), m_X(1u, 0.0) {}
+
+    double operator()(double x) const {
+        m_X[0] = x;
+        double lowerBound, upperBound;
+        if (m_Prior->minusLogJointCdf(m_X, m_Weights, lowerBound, upperBound) == false) {
+            throw std::runtime_error("Unable to compute c.d.f. at " +
+                                     core::CStringUtils::typeToString(x));
+        }
+        switch (m_Style) {
+        case E_Lower:
+            return -lowerBound;
+        case E_Upper:
+            return -upperBound;
+        case E_Mean:
+            return -(lowerBound + upperBound) / 2.0;
+        }
+        return -(lowerBound + upperBound) / 2.0;
+    }
+
+private:
+    EStyle m_Style;
+    const CMultimodalPrior* m_Prior;
+    TDoubleWeightsAry1Vec m_Weights;
+    //! Avoids creating the vector argument to minusLogJointCdf
+    //! more than once.
+    mutable TDouble1Vec m_X;
+};
 
 const std::size_t MODE_SPLIT_NUMBER_SAMPLES(50u);
 const std::size_t MODE_MERGE_NUMBER_SAMPLES(25u);
@@ -101,7 +168,7 @@ CMultimodalPrior::CMultimodalPrior(maths_t::EDataType dataType,
 CMultimodalPrior::CMultimodalPrior(maths_t::EDataType dataType, double decayRate, TPriorPtrVec& priors)
     : CPrior(dataType, decayRate) {
     m_Modes.reserve(priors.size());
-    for (std::size_t i = 0u; i < priors.size(); ++i) {
+    for (std::size_t i = 0; i < priors.size(); ++i) {
         m_Modes.emplace_back(i, std::move(priors[i]));
     }
 }
@@ -237,11 +304,10 @@ bool CMultimodalPrior::needsOffset() const {
 
 double CMultimodalPrior::adjustOffset(const TDouble1Vec& samples,
                                       const TDoubleWeightsAry1Vec& weights) {
-    double result = 0.0;
-
+    double result{0.0};
     if (this->needsOffset()) {
         TSizeDoublePr2Vec clusters;
-        for (std::size_t i = 0u; i < samples.size(); ++i) {
+        for (std::size_t i = 0; i < samples.size(); ++i) {
             m_Clusterer->cluster(samples[i], clusters);
             for (const auto& cluster : clusters) {
                 auto j = std::find_if(m_Modes.begin(), m_Modes.end(),
@@ -252,12 +318,11 @@ double CMultimodalPrior::adjustOffset(const TDouble1Vec& samples,
             }
         }
     }
-
     return result;
 }
 
 double CMultimodalPrior::offset() const {
-    double offset = 0.0;
+    double offset{0.0};
     for (const auto& mode : m_Modes) {
         offset = std::max(offset, mode.s_Prior->offset());
     }
@@ -305,13 +370,13 @@ void CMultimodalPrior::addSamples(const TDouble1Vec& samples,
     TSizeDoublePr2Vec clusters;
 
     try {
-        bool hasSeasonalScale = !this->isNonInformative() &&
-                                maths_t::hasSeasonalVarianceScale(weights);
-        double mean = hasSeasonalScale ? this->marginalLikelihoodMean() : 0.0;
+        bool hasSeasonalScale{this->isNonInformative() == false &&
+                              maths_t::hasSeasonalVarianceScale(weights)};
+        double mean{hasSeasonalScale ? this->marginalLikelihoodMean() : 0.0};
 
-        for (std::size_t i = 0u; i < samples.size(); ++i) {
-            double x = samples[i];
-            if (!CMathsFuncs::isFinite(x)) {
+        for (std::size_t i = 0; i < samples.size(); ++i) {
+            double x{samples[i]};
+            if (CMathsFuncs::isFinite(x) == false) {
                 LOG_ERROR(<< "Discarding " << x);
                 continue;
             }
@@ -327,13 +392,11 @@ void CMultimodalPrior::addSamples(const TDouble1Vec& samples,
             clusters.clear();
             m_Clusterer->add(x, clusters, maths_t::count(weight[0]));
 
-            auto addModeWeight = [](double sum, const TMode& mode) {
-                return sum + mode.weight();
-            };
-            double Z = std::accumulate(m_Modes.begin(), m_Modes.end(),
-                                       maths_t::count(weight[0]), addModeWeight);
+            double Z{std::accumulate(
+                m_Modes.begin(), m_Modes.end(), maths_t::count(weight[0]),
+                [](double sum, const TMode& mode) { return sum + mode.weight(); })};
 
-            double n = 0.0;
+            double n{0.0};
             for (const auto& cluster : clusters) {
                 auto k = std::find_if(m_Modes.begin(), m_Modes.end(),
                                       CSetTools::CIndexInSet(cluster.first));
@@ -360,7 +423,7 @@ void CMultimodalPrior::addSamples(const TDouble1Vec& samples,
 }
 
 void CMultimodalPrior::propagateForwardsByTime(double time) {
-    if (!CMathsFuncs::isFinite(time) || time < 0.0) {
+    if (CMathsFuncs::isFinite(time) == false || time < 0.0) {
         LOG_ERROR(<< "Bad propagation time " << time);
         return;
     }
@@ -386,11 +449,44 @@ void CMultimodalPrior::propagateForwardsByTime(double time) {
 }
 
 TDoubleDoublePr CMultimodalPrior::marginalLikelihoodSupport() const {
-    return CMultimodalPriorUtils::marginalLikelihoodSupport(m_Modes);
+    if (m_Modes.size() == 0) {
+        return {-std::numeric_limits<double>::max(), std::numeric_limits<double>::max()};
+    }
+    if (m_Modes.size() == 1) {
+        return m_Modes[0].s_Prior->marginalLikelihoodSupport();
+    }
+
+    TDoubleDoublePr result{-std::numeric_limits<double>::max(),
+                           std::numeric_limits<double>::max()};
+
+    // We define this is as the union of the mode supports.
+    for (const auto& mode : m_Modes) {
+        TDoubleDoublePr s{mode.s_Prior->marginalLikelihoodSupport()};
+        result.first = std::min(result.first, s.first);
+        result.second = std::max(result.second, s.second);
+    }
+
+    return result;
 }
 
 double CMultimodalPrior::marginalLikelihoodMean() const {
-    return CMultimodalPriorUtils::marginalLikelihoodMean(m_Modes);
+    if (m_Modes.size() == 0) {
+        return 0.0;
+    }
+    if (m_Modes.size() == 1) {
+        return m_Modes[0].s_Prior->marginalLikelihoodMean();
+    }
+
+    // By linearity we have that:
+    //   Integral{ x * Sum_i{ w(i) * f(x | i) } }
+    //     = Sum_i{ w(i) * Integral{ x * f(x | i) } }
+    //     = Sum_i{ w(i) * mean(i) }
+
+    TMeanAccumulator result;
+    for (const auto& mode : m_Modes) {
+        result.add(mode.s_Prior->marginalLikelihoodMean(), mode.weight());
+    }
+    return CBasicStatistics::mean(result);
 }
 
 double CMultimodalPrior::nearestMarginalLikelihoodMean(double value) const {
@@ -399,10 +495,10 @@ double CMultimodalPrior::nearestMarginalLikelihoodMean(double value) const {
         return 0.0;
     }
 
-    double mean = m_Modes[0].s_Prior->marginalLikelihoodMean();
-    double distance = std::fabs(value - mean);
-    double result = mean;
-    for (std::size_t i = 1u; i < m_Modes.size(); ++i) {
+    double mean{m_Modes[0].s_Prior->marginalLikelihoodMean()};
+    double distance{std::fabs(value - mean)};
+    double result{mean};
+    for (std::size_t i = 1; i < m_Modes.size(); ++i) {
         mean = m_Modes[i].s_Prior->marginalLikelihoodMean();
         if (std::fabs(value - mean) < distance) {
             distance = std::fabs(value - mean);
@@ -413,27 +509,156 @@ double CMultimodalPrior::nearestMarginalLikelihoodMean(double value) const {
 }
 
 double CMultimodalPrior::marginalLikelihoodMode(const TDoubleWeightsAry& weights) const {
-    return CMultimodalPriorUtils::marginalLikelihoodMode(m_Modes, weights);
+    if (m_Modes.size() == 0) {
+        return 0.0;
+    }
+    if (m_Modes.size() == 1) {
+        return m_Modes[0].s_Prior->marginalLikelihoodMode(weights);
+    }
+
+    using TMaxAccumulator = CBasicStatistics::SMax<double>::TAccumulator;
+
+    // We'll approximate this as the maximum likelihood mode (mode).
+    double result{0.0};
+
+    double seasonalScale{std::sqrt(maths_t::seasonalVarianceScale(weights))};
+    double countVarianceScale{maths_t::countVarianceScale(weights)};
+
+    // Declared outside the loop to minimize number of times they
+    // are created.
+    TDouble1Vec distributionMode(1);
+    TDoubleWeightsAry1Vec weight{maths_t::countVarianceScaleWeight(countVarianceScale)};
+
+    TMaxAccumulator maxLikelihood;
+    for (const auto& mode : m_Modes) {
+        double w{mode.weight()};
+        const auto& prior = mode.s_Prior;
+        distributionMode[0] = prior->marginalLikelihoodMode(weight[0]);
+        double likelihood;
+        if (prior->jointLogMarginalLikelihood(distributionMode, weight, likelihood) &
+            (maths_t::E_FpFailed | maths_t::E_FpOverflowed)) {
+            continue;
+        }
+        if (maxLikelihood.add(std::log(w) + likelihood)) {
+            result = distributionMode[0];
+        }
+    }
+
+    if (maths_t::hasSeasonalVarianceScale(weights)) {
+        double mean{this->marginalLikelihoodMean()};
+        result = mean + seasonalScale * (result - mean);
+    }
+
+    return result;
 }
 
 CMultimodalPrior::TDouble1Vec
 CMultimodalPrior::marginalLikelihoodModes(const TDoubleWeightsAry& weights) const {
     TDouble1Vec result(m_Modes.size());
-    for (std::size_t i = 0u; i < m_Modes.size(); ++i) {
+    for (std::size_t i = 0; i < m_Modes.size(); ++i) {
         result[i] = m_Modes[i].s_Prior->marginalLikelihoodMode(weights);
     }
     return result;
 }
 
 double CMultimodalPrior::marginalLikelihoodVariance(const TDoubleWeightsAry& weights) const {
-    return CMultimodalPriorUtils::marginalLikelihoodVariance(m_Modes, weights);
+    if (m_Modes.size() == 0) {
+        return std::numeric_limits<double>::max();
+    }
+    if (m_Modes.size() == 1) {
+        return m_Modes[0].s_Prior->marginalLikelihoodVariance(weights);
+    }
+
+    // By linearity we have that:
+    //   Integral{ (x - m)^2 * Sum_i{ w(i) * f(x | i) } }
+    //     = Sum_i{ w(i) * (Integral{ x^2 * f(x | i) } - m^2) }
+    //     = Sum_i{ w(i) * ((mi^2 + vi) - m^2) }
+
+    double varianceScale{maths_t::seasonalVarianceScale(weights) *
+                         maths_t::countVarianceScale(weights)};
+    double mean{this->marginalLikelihoodMean()};
+
+    TMeanAccumulator result;
+    for (const auto& mode : m_Modes) {
+        double w{mode.weight()};
+        double mm{mode.s_Prior->marginalLikelihoodMean()};
+        double mv{mode.s_Prior->marginalLikelihoodVariance()};
+        result.add((mm - mean) * (mm + mean) + mv, w);
+    }
+
+    return std::max(varianceScale * CBasicStatistics::mean(result), 0.0);
 }
 
 TDoubleDoublePr
 CMultimodalPrior::marginalLikelihoodConfidenceInterval(double percentage,
                                                        const TDoubleWeightsAry& weights) const {
-    return CMultimodalPriorUtils::marginalLikelihoodConfidenceInterval(
-        *this, m_Modes, percentage, weights);
+
+    TDoubleDoublePr support{this->marginalLikelihoodSupport()};
+
+    if (this->isNonInformative()) {
+        return support;
+    }
+
+    if (m_Modes.size() == 1) {
+        return m_Modes[0].s_Prior->marginalLikelihoodConfidenceInterval(percentage, weights);
+    }
+
+    percentage = CTools::truncate(percentage, 0.0, 100.0);
+    if (percentage == 100.0) {
+        return support;
+    }
+
+    using TMinMaxAccumulator = CBasicStatistics::CMinMax<double>;
+
+    // The multimodal distribution confidence interval must lie between the
+    // minimum and maximum of the corresponding modes' confidence intervals.
+    TMinMaxAccumulator lower;
+    TMinMaxAccumulator upper;
+    for (const auto& mode : m_Modes) {
+        auto interval = mode.s_Prior->marginalLikelihoodConfidenceInterval(percentage, weights);
+        lower.add(interval.first);
+        upper.add(interval.second);
+    }
+
+    percentage /= 100.0;
+    double p1{maths_t::count(weights) * std::log((1.0 - percentage) / 2.0)};
+    double p2{maths_t::count(weights) * std::log((1.0 + percentage) / 2.0)};
+
+    double vs{maths_t::seasonalVarianceScale(weights)};
+    double mean{vs != 1.0 ? this->marginalLikelihoodMean() : 0.0};
+
+    CLogCdf fl{CLogCdf::E_Lower, *this, weights};
+    CLogCdf fu{CLogCdf::E_Upper, *this, weights};
+
+    auto computePercentile = [&](const CLogCdf& f, double p, const TMinMaxAccumulator& bounds) {
+        auto fMinusp = [&f, p](double x) { return f(x) - p; };
+        double a{bounds.min() < mean ? mean - std::max(vs, 1.0) * (mean - bounds.min())
+                                     : bounds.min()};
+        double b{bounds.max() > mean ? mean + std::max(vs, 1.0) * (bounds.max() - mean)
+                                     : bounds.max()};
+        double fa{fMinusp(a)};
+        double fb{fMinusp(b)};
+        std::size_t maxIterations{25};
+        CEqualWithTolerance<double> equal{CToleranceTypes::E_AbsoluteTolerance,
+                                          std::min(std::numeric_limits<double>::epsilon() * b,
+                                                   1e-3 * p / std::max(fa, fb))};
+        double percentile;
+        CSolvers::solve(a, b, fa, fb, fMinusp, maxIterations, equal, percentile);
+        LOG_TRACE(<< "p1 = " << p << ", x = " << percentile << ", f(x) = " << f(percentile)
+                  << " brackets = [" << a << "," << b << "]");
+        return percentile;
+    };
+
+    TDoubleDoublePr result{mean - vs * (mean - lower.min()),
+                           mean + vs * (upper.max() - mean)};
+    try {
+        result.first = computePercentile(fl, p1, lower);
+        result.second = computePercentile(fu, p2, upper);
+    } catch (const std::exception& e) {
+        LOG_ERROR(<< "Unable to find left percentile: " << e.what()
+                  << ", percentiles = [" << p1 << "," << p2 << "]");
+    }
+    return result;
 }
 
 maths_t::EFloatingPointErrorStatus
@@ -462,14 +687,129 @@ CMultimodalPrior::jointLogMarginalLikelihood(const TDouble1Vec& samples,
         // underflow and pollute the floating point environment. This
         // may cause issues for some library function implementations
         // (see fe*exceptflag for more details).
-        result = boost::numeric::bounds<double>::lowest();
+        result = -std::numeric_limits<double>::max();
         return maths_t::E_FpOverflowed;
     }
 
-    return m_Modes.size() == 1
-               ? m_Modes[0].s_Prior->jointLogMarginalLikelihood(samples, weights, result)
-               : CMultimodalPriorUtils::jointLogMarginalLikelihood(
-                     m_Modes, samples, weights, result);
+    if (m_Modes.size() == 1) {
+        return m_Modes[0].s_Prior->jointLogMarginalLikelihood(samples, weights, result);
+    }
+
+    // The likelihood can be computed from the conditional likelihood
+    // that a sample is from each mode. In particular, the likelihood
+    // of a sample x is:
+    //   L(x) = Sum_m{ L(x | m) * p(m) }
+    //
+    // where,
+    //   L(x | m) is the likelihood the sample is from the m'th mode,
+    //   p(m) is the probability a sample is from the m'th mode.
+    //
+    // We compute the combined likelihood by taking the product of the
+    // individual likelihoods. Note, this brushes over the fact that the
+    // joint marginal likelihood that a collection of samples is from
+    // the i'th mode is not just the product of the likelihoods that the
+    // individual samples are from the i'th mode since we're integrating
+    // over a prior. Really, we should compute likelihoods over all
+    // possible assignments of the samples to the modes and use the fact
+    // that:
+    //   P(a) = Product_i{ Sum_m{ p(m) * I{a(i) = m} } }
+    //
+    // where,
+    //   P(a) is the probability of a given assignment,
+    //   p(m) is the probability a sample is from the m'th mode,
+    //   I{.} is the indicator function.
+    //
+    // The approximation is increasingly accurate as the prior distribution
+    // on each mode narrows.
+
+    result = 0.0;
+
+    // Declared outside the loop to minimize number of times it is created.
+    TDouble1Vec sample(1);
+    TSizeDoublePr5Vec modeLogLikelihoods;
+    modeLogLikelihoods.reserve(m_Modes.size());
+
+    double mean{maths_t::hasSeasonalVarianceScale(weights) ? this->marginalLikelihoodMean()
+                                                           : 0.0};
+    TDoubleWeightsAry1Vec weight{TWeights::UNIT};
+    try {
+        for (std::size_t i = 0; i < samples.size(); ++i) {
+            double n{maths_t::countForUpdate(weights[i])};
+            double seasonalScale{std::sqrt(maths_t::seasonalVarianceScale(weights[i]))};
+            double logSeasonalScale{seasonalScale != 1.0 ? std::log(seasonalScale) : 0.0};
+
+            sample[0] = mean + (samples[i] - mean) / seasonalScale;
+            maths_t::setCountVarianceScale(maths_t::countVarianceScale(weights[i]),
+                                           weight[0]);
+
+            // We re-normalize so that the maximum log likelihood is one
+            // to avoid underflow.
+            modeLogLikelihoods.clear();
+            double maxLogLikelihood{-std::numeric_limits<double>::max()};
+
+            for (std::size_t j = 0; j < m_Modes.size(); ++j) {
+                double modeLogLikelihood;
+                maths_t::EFloatingPointErrorStatus status{m_Modes[j].s_Prior->jointLogMarginalLikelihood(
+                    sample, weight, modeLogLikelihood)};
+                if (status & maths_t::E_FpFailed) {
+                    // Logging handled at a lower level.
+                    return status;
+                }
+                if ((status & maths_t::E_FpOverflowed) == false) {
+                    modeLogLikelihoods.emplace_back(j, modeLogLikelihood);
+                    maxLogLikelihood = std::max(maxLogLikelihood, modeLogLikelihood);
+                }
+            }
+
+            if (modeLogLikelihoods.empty()) {
+                // Technically, the marginal likelihood is zero here
+                // so the log would be infinite. We use minus max
+                // double because log(0) = HUGE_VALUE, which causes
+                // problems for Windows. Calling code is notified
+                // when the calculation overflows and should avoid
+                // taking the exponential since this will underflow
+                // and pollute the floating point environment. This
+                // may cause issues for some library function
+                // implementations (see fe*exceptflag for more details).
+                result = -std::numeric_limits<double>::max();
+                return maths_t::E_FpOverflowed;
+            }
+
+            LOG_TRACE(<< "modeLogLikelihoods = "
+                      << core::CContainerPrinter::print(modeLogLikelihoods));
+
+            double sampleLikelihood{0.0};
+            double Z{0.0};
+
+            for (const auto& modeLogLikelihood : modeLogLikelihoods) {
+                double w{m_Modes[modeLogLikelihood.first].weight()};
+                // Divide through by the largest value to avoid underflow.
+                sampleLikelihood += w * std::exp(modeLogLikelihood.second - maxLogLikelihood);
+                Z += w;
+            }
+
+            sampleLikelihood /= Z;
+            double sampleLogLikelihood{n * (std::log(sampleLikelihood) + maxLogLikelihood)};
+
+            LOG_TRACE(<< "sample = " << core::CContainerPrinter::print(sample)
+                      << ", maxLogLikelihood = " << maxLogLikelihood
+                      << ", sampleLogLikelihood = " << sampleLogLikelihood);
+
+            result += sampleLogLikelihood - n * logSeasonalScale;
+        }
+    } catch (const std::exception& e) {
+        LOG_ERROR(<< "Failed to compute likelihood: " << e.what());
+        return maths_t::E_FpFailed;
+    }
+
+    maths_t::EFloatingPointErrorStatus status{CMathsFuncs::fpStatus(result)};
+    if (status & maths_t::E_FpFailed) {
+        LOG_ERROR(<< "Failed to compute likelihood (" << this->debugWeights() << ")");
+        LOG_ERROR(<< "samples = " << core::CContainerPrinter::print(samples));
+        LOG_ERROR(<< "weights = " << core::CContainerPrinter::print(weights));
+    }
+    LOG_TRACE(<< "Joint log likelihood = " << result);
+    return status;
 }
 
 void CMultimodalPrior::sampleMarginalLikelihood(std::size_t numberSamples,
@@ -481,23 +821,63 @@ void CMultimodalPrior::sampleMarginalLikelihood(std::size_t numberSamples,
         return;
     }
 
-    CMultimodalPriorUtils::sampleMarginalLikelihood(m_Modes, numberSamples, samples);
+    samples.clear();
+
+    if (m_Modes.size() == 1) {
+        m_Modes[0].s_Prior->sampleMarginalLikelihood(numberSamples, samples);
+        return;
+    }
+
+    // We sample each mode according to its weight.
+
+    TDoubleVec normalizedWeights;
+    normalizedWeights.reserve(m_Modes.size());
+    double Z{0.0};
+
+    for (const auto& mode : m_Modes) {
+        double weight{mode.weight()};
+        normalizedWeights.push_back(weight);
+        Z += weight;
+    }
+    for (auto& weight : normalizedWeights) {
+        weight /= Z;
+    }
+
+    CSampling::TSizeVec sampling;
+    CSampling::weightedSample(numberSamples, normalizedWeights, sampling);
+    LOG_TRACE(<< "normalizedWeights = " << core::CContainerPrinter::print(normalizedWeights)
+              << ", sampling = " << core::CContainerPrinter::print(sampling));
+
+    if (sampling.size() != m_Modes.size()) {
+        LOG_ERROR(<< "Failed to sample marginal likelihood");
+        return;
+    }
+
+    samples.reserve(numberSamples);
+    TDouble1Vec modeSamples;
+    for (std::size_t i = 0; i < m_Modes.size(); ++i) {
+        m_Modes[i].s_Prior->sampleMarginalLikelihood(sampling[i], modeSamples);
+        LOG_TRACE(<< "modeSamples = " << core::CContainerPrinter::print(modeSamples));
+        std::copy(modeSamples.begin(), modeSamples.end(), std::back_inserter(samples));
+    }
+    LOG_TRACE(<< "samples = " << core::CContainerPrinter::print(samples));
 }
 
 bool CMultimodalPrior::minusLogJointCdf(const TDouble1Vec& samples,
                                         const TDoubleWeightsAry1Vec& weights,
                                         double& lowerBound,
                                         double& upperBound) const {
-    return CMultimodalPriorUtils::minusLogJointCdf(m_Modes, samples, weights,
-                                                   lowerBound, upperBound);
+    return this->minusLogJointCdfImpl(CMinusLogJointCdf{}, samples, weights,
+                                      lowerBound, upperBound);
 }
 
 bool CMultimodalPrior::minusLogJointCdfComplement(const TDouble1Vec& samples,
                                                   const TDoubleWeightsAry1Vec& weights,
                                                   double& lowerBound,
                                                   double& upperBound) const {
-    return CMultimodalPriorUtils::minusLogJointCdfComplement(
-        m_Modes, samples, weights, lowerBound, upperBound);
+
+    return this->minusLogJointCdfImpl(CMinusLogJointCdfComplement{}, samples,
+                                      weights, lowerBound, upperBound);
 }
 
 bool CMultimodalPrior::probabilityOfLessLikelySamples(maths_t::EProbabilityCalculation calculation,
@@ -506,23 +886,235 @@ bool CMultimodalPrior::probabilityOfLessLikelySamples(maths_t::EProbabilityCalcu
                                                       double& lowerBound,
                                                       double& upperBound,
                                                       maths_t::ETail& tail) const {
-    return CMultimodalPriorUtils::probabilityOfLessLikelySamples(
-        *this, m_Modes, calculation, samples, weights, lowerBound, upperBound, tail);
+    lowerBound = upperBound = 1.0;
+    tail = maths_t::E_UndeterminedTail;
+
+    if (samples.empty()) {
+        LOG_ERROR(<< "Can't compute distribution for empty sample set");
+        return false;
+    }
+    if (this->isNonInformative()) {
+        return true;
+    }
+
+    if (m_Modes.size() == 1) {
+        return m_Modes[0].s_Prior->probabilityOfLessLikelySamples(
+            calculation, samples, weights, lowerBound, upperBound, tail);
+    }
+
+    // Ideally we'd find the probability of the set of samples whose
+    // total likelihood is less than or equal to that of the specified
+    // samples, i.e. the probability of the set
+    //   R = { y | L(y) < L(x) }
+    //
+    // where,
+    //   x = {x(1), x(2), ..., x(n)} is the sample vector.
+    //   y is understood to be a vector quantity.
+    //
+    // This is not *trivially* related to the probability that the
+    // probabilities of the sets
+    //   R(i) = { y | L(y) < L(x(i)) }
+    //
+    // since the joint conditional likelihood must be integrated over
+    // priors for the parameters. However, we'll approximate this as
+    // the joint probability (of a collection of standard normal R.Vs.)
+    // having probabilities {P(R(i))}. This becomes increasingly accurate
+    // as the prior distribution narrows.
+    //
+    // For the two sided calculation, we use the fact that the likelihood
+    // function decreases monotonically away from the interval [a, b]
+    // whose end points are the leftmost and rightmost modes' modes
+    // since all component likelihoods decrease away from this interval.
+    //
+    // To evaluate the probability in the interval [a, b] we relax
+    // the hard constraint that regions where f > f(x) contribute
+    // zero probability. In particular, we note that we can write
+    // the probability as:
+    //   P = Integral{ I(f(s) < f(x)) * f(s) }ds
+    //
+    // and that:
+    //   I(f(s) < f(x)) = lim_{k->inf}{ exp(-k * (f(s)/f(x) - 1))
+    //                                  / (1 + exp(-k * (f(s)/f(x) - 1))) }
+    //
+    // We evaluate a smoother integral, i.e. smaller p, initially
+    // to find out which regions contribute the most to P and then
+    // re-evaluate those regions we need with higher resolution
+    // using the fact that the maximum error in the approximation
+    // of I(f(s) < f(x)) is 0.5.
+
+    switch (calculation) {
+    case maths_t::E_OneSidedBelow:
+        if (this->minusLogJointCdf(samples, weights, upperBound, lowerBound) == false) {
+            LOG_ERROR(<< "Failed computing probability of less likely samples: "
+                      << core::CContainerPrinter::print(samples));
+            return false;
+        }
+        lowerBound = std::exp(-lowerBound);
+        upperBound = std::exp(-upperBound);
+        tail = maths_t::E_LeftTail;
+        break;
+
+    case maths_t::E_TwoSided: {
+        static const double EPS{1000.0 * std::numeric_limits<double>::epsilon()};
+        static const std::size_t MAX_ITERATIONS{20};
+
+        CJointProbabilityOfLessLikelySamples lowerBoundCalculator;
+        CJointProbabilityOfLessLikelySamples upperBoundCalculator;
+
+        TDoubleDoublePr support{this->marginalLikelihoodSupport()};
+        support.first = (1.0 + (support.first > 0.0 ? EPS : -EPS)) * support.first;
+        support.second = (1.0 + (support.first > 0.0 ? EPS : -EPS)) * support.second;
+        bool hasSeasonalScale{maths_t::hasSeasonalVarianceScale(weights)};
+        double mean{hasSeasonalScale ? this->marginalLikelihoodMean() : 0.0};
+
+        double a{std::numeric_limits<double>::max()};
+        double b{-std::numeric_limits<double>::max()};
+        double Z{0.0};
+        for (const auto& mode : m_Modes) {
+            double m{mode.s_Prior->marginalLikelihoodMode()};
+            a = std::min(a, m);
+            b = std::max(b, m);
+            Z += mode.weight();
+        }
+        a = CTools::truncate(a, support.first, support.second);
+        b = CTools::truncate(b, support.first, support.second);
+        LOG_TRACE(<< "a = " << a << ", b = " << b << ", Z = " << Z);
+
+        // Declared outside the loop to minimize the number of times
+        // it is created.
+        TDoubleWeightsAry1Vec weight(1);
+
+        int tail_{0};
+        for (std::size_t i = 0; i < samples.size(); ++i) {
+            double x{samples[i]};
+            weight[0] = weights[i];
+            if (hasSeasonalScale) {
+                x = mean + (x - mean) /
+                               std::sqrt(maths_t::seasonalVarianceScale(weight[0]));
+                maths_t::setSeasonalVarianceScale(1.0, weight[0]);
+            }
+
+            double fx;
+            maths_t::EFloatingPointErrorStatus status =
+                this->jointLogMarginalLikelihood({x}, weight, fx);
+            if (status & maths_t::E_FpFailed) {
+                LOG_ERROR(<< "Unable to compute likelihood for " << x);
+                return false;
+            }
+            if (status & maths_t::E_FpOverflowed) {
+                lowerBound = upperBound = 0.0;
+                return true;
+            }
+            LOG_TRACE(<< "x = " << x << ", f(x) = " << fx);
+
+            CPrior::CLogMarginalLikelihood logLikelihood{*this, weight};
+
+            CTools::CMixtureProbabilityOfLessLikelySample calculator{m_Modes.size(),
+                                                                     x, fx, a, b};
+            for (const auto& mode : m_Modes) {
+                double w{mode.weight() / Z};
+                double centre{mode.s_Prior->marginalLikelihoodMode(weight[0])};
+                double spread{
+                    std::sqrt(mode.s_Prior->marginalLikelihoodVariance(weight[0]))};
+                calculator.addMode(w, centre, spread);
+                tail_ = tail_ | (x < centre ? maths_t::E_LeftTail : maths_t::E_RightTail);
+            }
+
+            double sampleLowerBound{0.0};
+            double sampleUpperBound{0.0};
+
+            double lb, ub;
+
+            double xl;
+            CEqualWithTolerance<double> lequal{CToleranceTypes::E_AbsoluteTolerance,
+                                               EPS * a};
+            if (calculator.leftTail(logLikelihood, MAX_ITERATIONS, lequal, xl)) {
+                this->minusLogJointCdf({xl}, weight, lb, ub);
+                sampleLowerBound += std::exp(std::min(-lb, -ub));
+                sampleUpperBound += std::exp(std::max(-lb, -ub));
+            } else {
+                this->minusLogJointCdf({xl}, weight, lb, ub);
+                sampleUpperBound += std::exp(std::max(-lb, -ub));
+            }
+
+            double xr;
+            CEqualWithTolerance<double> requal{CToleranceTypes::E_AbsoluteTolerance,
+                                               EPS * b};
+            if (calculator.rightTail(logLikelihood, MAX_ITERATIONS, requal, xr)) {
+                this->minusLogJointCdfComplement({xr}, weight, lb, ub);
+                sampleLowerBound += std::exp(std::min(-lb, -ub));
+                sampleUpperBound += std::exp(std::max(-lb, -ub));
+            } else {
+                this->minusLogJointCdfComplement({xr}, weight, lb, ub);
+                sampleUpperBound += std::exp(std::max(-lb, -ub));
+            }
+
+            double p{0.0};
+            if (a < b) {
+                p = calculator.calculate(logLikelihood, sampleLowerBound);
+            }
+
+            LOG_TRACE(<< "sampleLowerBound = " << sampleLowerBound
+                      << ", sampleUpperBound = " << sampleUpperBound << " p = " << p);
+
+            lowerBoundCalculator.add(CTools::truncate(sampleLowerBound + p, 0.0, 1.0));
+            upperBoundCalculator.add(CTools::truncate(sampleUpperBound + p, 0.0, 1.0));
+        }
+
+        if (lowerBoundCalculator.calculate(lowerBound) == false ||
+            upperBoundCalculator.calculate(upperBound) == false) {
+            LOG_ERROR(<< "Couldn't compute probability of less likely samples:"
+                      << " " << lowerBoundCalculator << " " << upperBoundCalculator);
+            return false;
+        }
+        tail = static_cast<maths_t::ETail>(tail_);
+    } break;
+
+    case maths_t::E_OneSidedAbove:
+        if (this->minusLogJointCdfComplement(samples, weights, upperBound, lowerBound) == false) {
+            LOG_ERROR(<< "Failed computing probability of less likely samples: "
+                      << core::CContainerPrinter::print(samples));
+            return false;
+        }
+        lowerBound = std::exp(-lowerBound);
+        upperBound = std::exp(-upperBound);
+        tail = maths_t::E_RightTail;
+        break;
+    }
+
+    return true;
 }
 
 bool CMultimodalPrior::isNonInformative() const {
-    return CMultimodalPriorUtils::isNonInformative(m_Modes);
+    return m_Modes.empty() ||
+           (m_Modes.size() == 1 && m_Modes[0].s_Prior->isNonInformative());
 }
 
 void CMultimodalPrior::print(const std::string& indent, std::string& result) const {
-    CMultimodalPriorUtils::print(m_Modes, indent, result);
+    result += "\n" + indent + "multimodal";
+    if (this->isNonInformative()) {
+        result += " non-informative";
+        return;
+    }
+
+    double Z{0.0};
+    for (const auto& mode : m_Modes) {
+        Z += mode.weight();
+    }
+    result += ":";
+    for (const auto& mode : m_Modes) {
+        double weight{mode.weight() / Z};
+        std::string indent_{indent + " weight " +
+                            core::CStringUtils::typeToStringPretty(weight) + "  "};
+        mode.s_Prior->print(indent_, result);
+    }
 }
 
 std::string CMultimodalPrior::printJointDensityFunction() const {
     return "Not supported";
 }
 
-uint64_t CMultimodalPrior::checksum(uint64_t seed) const {
+std::uint64_t CMultimodalPrior::checksum(std::uint64_t seed) const {
     seed = this->CPrior::checksum(seed);
     seed = CChecksum::calculate(seed, m_Clusterer);
     seed = CChecksum::calculate(seed, m_SeedPrior);
@@ -554,7 +1146,7 @@ void CMultimodalPrior::acceptPersistInserter(core::CStatePersistInserter& insert
     inserter.insertLevel(SEED_PRIOR_TAG, std::bind<void>(CPriorStateSerialiser(),
                                                          std::cref(*m_SeedPrior),
                                                          std::placeholders::_1));
-    for (std::size_t i = 0u; i < m_Modes.size(); ++i) {
+    for (std::size_t i = 0; i < m_Modes.size(); ++i) {
         inserter.insertLevel(MODE_TAG, std::bind(&TMode::acceptPersistInserter,
                                                  &m_Modes[i], std::placeholders::_1));
     }
@@ -569,7 +1161,7 @@ std::size_t CMultimodalPrior::numberModes() const {
 
 bool CMultimodalPrior::checkInvariants(const std::string& tag) const {
 
-    bool result = true;
+    bool result{true};
 
     if (m_Modes.size() != m_Clusterer->numberClusters()) {
         LOG_ERROR(<< tag << "# modes = " << m_Modes.size()
@@ -577,19 +1169,19 @@ bool CMultimodalPrior::checkInvariants(const std::string& tag) const {
         result = false;
     }
 
-    double numberSamples = this->numberSamples();
-    double modeSamples = 0.0;
+    double numberSamples{this->numberSamples()};
+    double modeSamples{0.0};
     for (const auto& mode : m_Modes) {
-        if (!m_Clusterer->hasCluster(mode.s_Index)) {
+        if (m_Clusterer->hasCluster(mode.s_Index) == false) {
             LOG_ERROR(<< tag << "Expected cluster for mode = " << mode.s_Index);
             result = false;
         }
         modeSamples += mode.s_Prior->numberSamples();
     }
 
-    CEqualWithTolerance<double> equal(
-        CToleranceTypes::E_AbsoluteTolerance | CToleranceTypes::E_RelativeTolerance, 1e-3);
-    if (!equal(modeSamples, numberSamples)) {
+    CEqualWithTolerance<double> equal{
+        CToleranceTypes::E_AbsoluteTolerance | CToleranceTypes::E_RelativeTolerance, 1e-3};
+    if (equal(modeSamples, numberSamples) == false) {
         LOG_ERROR(<< tag << "Sum mode samples = " << modeSamples
                   << ", total samples = " << numberSamples);
         result = false;
@@ -604,6 +1196,114 @@ bool CMultimodalPrior::participatesInModelSelection() const {
 
 double CMultimodalPrior::unmarginalizedParameters() const {
     return std::max(static_cast<double>(m_Modes.size()), 1.0) - 1.0;
+}
+
+template<typename CDF>
+bool CMultimodalPrior::minusLogJointCdfImpl(CDF minusLogCdf,
+                                            const TDouble1Vec& samples,
+                                            const TDoubleWeightsAry1Vec& weights,
+                                            double& lowerBound,
+                                            double& upperBound) const {
+    lowerBound = upperBound = 0.0;
+
+    if (samples.empty()) {
+        LOG_ERROR(<< "Can't compute c.d.f. for empty sample set");
+        return false;
+    }
+
+    if (m_Modes.size() == 1) {
+        return minusLogCdf(m_Modes[0].s_Prior, samples, weights, lowerBound, upperBound);
+    }
+
+    using TMinAccumulator = CBasicStatistics::COrderStatisticsStack<double, 1>;
+
+    // The c.d.f. of the marginal likelihood is the weighted sum
+    // of the c.d.fs of each mode since:
+    //   cdf(x) = Integral{ L(u) }du
+    //          = Integral{ Sum_m{ L(u | m) p(m) } }du
+    //          = Sum_m{ Integral{ L(u | m) ) p(m) }du }
+
+    // Declared outside the loop to minimize the number of times
+    // they are created.
+    TDouble1Vec sample(1);
+    TDoubleWeightsAry1Vec weight{TWeights::UNIT};
+    TDoubleVec modeLowerBounds;
+    TDoubleVec modeUpperBounds;
+    modeLowerBounds.reserve(m_Modes.size());
+    modeUpperBounds.reserve(m_Modes.size());
+
+    try {
+        double mean{maths_t::hasSeasonalVarianceScale(weights) ? this->marginalLikelihoodMean()
+                                                               : 0.0};
+
+        for (std::size_t i = 0; i < samples.size(); ++i) {
+            double n{maths_t::count(weights[i])};
+            double seasonalScale{std::sqrt(maths_t::seasonalVarianceScale(weights[i]))};
+
+            if (this->isNonInformative()) {
+                lowerBound -= n * std::log(CTools::IMPROPER_CDF);
+                upperBound -= n * std::log(CTools::IMPROPER_CDF);
+                continue;
+            }
+
+            sample[0] = mean + (samples[i] - mean) / seasonalScale;
+            maths_t::setCountVarianceScale(maths_t::countVarianceScale(weights[i]),
+                                           weight[0]);
+
+            // We re-normalize so that the maximum log c.d.f. is one
+            // to avoid underflow.
+            TMinAccumulator minLowerBound;
+            TMinAccumulator minUpperBound;
+            modeLowerBounds.clear();
+            modeUpperBounds.clear();
+
+            for (const auto& mode : m_Modes) {
+                double modeLowerBound;
+                double modeUpperBound;
+                if (minusLogCdf(mode.s_Prior, sample, weight, modeLowerBound,
+                                modeUpperBound) == false) {
+                    LOG_ERROR(<< "Unable to compute c.d.f. for "
+                              << core::CContainerPrinter::print(samples));
+                    return false;
+                }
+                minLowerBound.add(modeLowerBound);
+                minUpperBound.add(modeUpperBound);
+                modeLowerBounds.push_back(modeLowerBound);
+                modeUpperBounds.push_back(modeUpperBound);
+            }
+
+            TMeanAccumulator sampleLowerBound;
+            TMeanAccumulator sampleUpperBound;
+
+            for (std::size_t j = 0; j < m_Modes.size(); ++j) {
+                LOG_TRACE(<< "Mode -log(c.d.f.) = [" << modeLowerBounds[j]
+                          << "," << modeUpperBounds[j] << "]");
+                double w{m_Modes[j].weight()};
+                // Divide through by the largest value to avoid underflow.
+                // Remember we are working with minus logs so the largest
+                // value corresponds to the smallest log.
+                sampleLowerBound.add(std::exp(-(modeLowerBounds[j] - minLowerBound[0])), w);
+                sampleUpperBound.add(std::exp(-(modeUpperBounds[j] - minUpperBound[0])), w);
+            }
+
+            lowerBound += n * std::max(minLowerBound[0] -
+                                           std::log(CBasicStatistics::mean(sampleLowerBound)),
+                                       0.0);
+            upperBound += n * std::max(minUpperBound[0] -
+                                           std::log(CBasicStatistics::mean(sampleUpperBound)),
+                                       0.0);
+
+            LOG_TRACE(<< "sample = " << core::CContainerPrinter::print(sample) << ", sample -log(c.d.f.) = ["
+                      << sampleLowerBound << "," << sampleUpperBound << "]");
+        }
+    } catch (const std::exception& e) {
+        LOG_ERROR(<< "Failed to calculate c.d.f.: " << e.what());
+        return false;
+    }
+
+    LOG_TRACE(<< "Joint -log(c.d.f.) = [" << lowerBound << "," << upperBound << "]");
+
+    return true;
 }
 
 std::string CMultimodalPrior::debugWeights() const {
@@ -626,11 +1326,11 @@ void CMultimodalPrior::CModeSplitCallback::operator()(std::size_t sourceIndex,
 
     // Remove the split mode.
     auto mode = std::find_if(modes.begin(), modes.end(), CSetTools::CIndexInSet(sourceIndex));
-    double numberSamples = mode != modes.end() ? mode->weight() : 0.0;
+    double numberSamples{mode != modes.end() ? mode->weight() : 0.0};
     modes.erase(mode);
 
-    double pLeft = m_Prior->m_Clusterer->probability(leftSplitIndex);
-    double pRight = m_Prior->m_Clusterer->probability(rightSplitIndex);
+    double pLeft{m_Prior->m_Clusterer->probability(leftSplitIndex)};
+    double pRight{m_Prior->m_Clusterer->probability(rightSplitIndex)};
     double Z = (pLeft + pRight);
     if (Z > 0.0) {
         pLeft /= Z;
@@ -645,14 +1345,15 @@ void CMultimodalPrior::CModeSplitCallback::operator()(std::size_t sourceIndex,
     modes.emplace_back(leftSplitIndex, TPriorPtr(m_Prior->m_SeedPrior->clone()));
     {
         TDoubleVec samples;
-        if (!m_Prior->m_Clusterer->sample(leftSplitIndex, MODE_SPLIT_NUMBER_SAMPLES, samples)) {
+        if (m_Prior->m_Clusterer->sample(leftSplitIndex, MODE_SPLIT_NUMBER_SAMPLES,
+                                         samples) == false) {
             LOG_ERROR(<< "Couldn't find cluster for " << leftSplitIndex);
         }
         LOG_TRACE(<< "samples = " << core::CContainerPrinter::print(samples));
 
-        double wl = pLeft * numberSamples;
-        double ws = std::min(wl, 4.0);
-        double n = static_cast<double>(samples.size());
+        double wl{pLeft * numberSamples};
+        double ws{std::min(wl, 4.0)};
+        double n{static_cast<double>(samples.size())};
         LOG_TRACE(<< "# left = " << wl);
 
         TDoubleWeightsAry1Vec weights(samples.size(), maths_t::countWeight(ws / n));
@@ -669,14 +1370,15 @@ void CMultimodalPrior::CModeSplitCallback::operator()(std::size_t sourceIndex,
     modes.emplace_back(rightSplitIndex, TPriorPtr(m_Prior->m_SeedPrior->clone()));
     {
         TDoubleVec samples;
-        if (!m_Prior->m_Clusterer->sample(rightSplitIndex, MODE_SPLIT_NUMBER_SAMPLES, samples)) {
+        if (m_Prior->m_Clusterer->sample(rightSplitIndex, MODE_SPLIT_NUMBER_SAMPLES,
+                                         samples) == false) {
             LOG_ERROR(<< "Couldn't find cluster for " << rightSplitIndex);
         }
         LOG_TRACE(<< "samples = " << core::CContainerPrinter::print(samples));
 
-        double wr = pRight * numberSamples;
-        double ws = std::min(wr, 4.0);
-        double n = static_cast<double>(samples.size());
+        double wr{pRight * numberSamples};
+        double ws{std::min(wr, 4.0)};
+        double n{static_cast<double>(samples.size())};
         LOG_TRACE(<< "# right = " << wr);
 
         TDoubleWeightsAry1Vec weights(samples.size(), maths_t::countWeight(ws / n));
@@ -689,7 +1391,7 @@ void CMultimodalPrior::CModeSplitCallback::operator()(std::size_t sourceIndex,
         }
     }
 
-    if (!m_Prior->checkInvariants("SPLIT: ")) {
+    if (m_Prior->checkInvariants("SPLIT: ") == false) {
         LOG_ERROR(<< "# samples = " << numberSamples << ", # modes = " << modes.size()
                   << ", pLeft = " << pLeft << ", pRight = " << pRight);
     }
@@ -709,16 +1411,16 @@ void CMultimodalPrior::CModeMergeCallback::operator()(std::size_t leftMergeIndex
 
     LOG_TRACE(<< "Merging modes with indices " << leftMergeIndex << " " << rightMergeIndex);
 
-    TModeVec& modes = m_Prior->m_Modes;
+    TModeVec& modes{m_Prior->m_Modes};
 
     // Create the new mode.
     TMode newMode(targetIndex, TPriorPtr(m_Prior->m_SeedPrior->clone()));
 
-    double wl = 0.0;
-    double wr = 0.0;
-    double w = 0.0;
-    std::size_t nl = 0;
-    std::size_t nr = 0;
+    double wl{0.0};
+    double wr{0.0};
+    double w{0.0};
+    std::size_t nl{0};
+    std::size_t nr{0};
     TDouble1Vec samples;
 
     auto leftMode = std::find_if(modes.begin(), modes.end(),
@@ -748,9 +1450,9 @@ void CMultimodalPrior::CModeMergeCallback::operator()(std::size_t leftMergeIndex
     }
 
     if (w > 0.0) {
-        double nl_ = static_cast<double>(nl);
-        double nr_ = static_cast<double>(nr);
-        double Z = (nl_ * wl + nr_ * wr) / (nl_ + nr_);
+        double nl_{static_cast<double>(nl)};
+        double nr_{static_cast<double>(nr)};
+        double Z{(nl_ * wl + nr_ * wr) / (nl_ + nr_)};
         wl /= Z;
         wr /= Z;
     }
@@ -758,8 +1460,8 @@ void CMultimodalPrior::CModeMergeCallback::operator()(std::size_t leftMergeIndex
     LOG_TRACE(<< "samples = " << core::CContainerPrinter::print(samples));
     LOG_TRACE(<< "w = " << w << ", wl = " << wl << ", wr = " << wr);
 
-    double ws = std::min(w, 4.0);
-    double n = static_cast<double>(samples.size());
+    double ws{std::min(w, 4.0)};
+    double n{static_cast<double>(samples.size())};
 
     TDoubleWeightsAry1Vec weights;
     weights.reserve(samples.size());
